+++ temporian/implementation/numpy_cc/operators/filter_moving_count.h
@@ -1,4 +1,4 @@
 #include <pybind11/numpy.h>
 #include <pybind11/pybind11.h>
 
-void init_filter_max_moving_count(pybind11::module &m);
+void init_filter_moving_count(pybind11::module &m);