+++ temporian/implementation/numpy_cc/operators/pyinit.cc
@@ -2,7 +2,7 @@
 #include <pybind11/pybind11.h>
 
 #include "temporian/implementation/numpy_cc/operators/add_index.h"
-#include "temporian/implementation/numpy_cc/operators/filter_max_moving_count.h"
+#include "temporian/implementation/numpy_cc/operators/filter_moving_count.h"
 #include "temporian/implementation/numpy_cc/operators/join.h"
 #include "temporian/implementation/numpy_cc/operators/resample.h"
 #include "temporian/implementation/numpy_cc/operators/since_last.h"
@@ -18,5 +18,5 @@ PYBIND11_MODULE(operators_cc, m) {
   init_window(m);
   init_join(m);
   init_add_index(m);
-  init_filter_max_moving_count(m);
+  init_filter_moving_count(m);
 }