+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- Add `filter_max_moving_count` operator.
+- Add `filter_moving_count` operator.
 
 ### Improvements